deployment/requirements.txt
aws-cdk-lib>=2.150.0
aws-cdk.aws-bedrock-agentcore-alpha>=2.150.0a0
aws-cdk.aws-bedrock-alpha>=2.150.0a0
constructs>=10.0.0
